backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pymongo>=4.6.1
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
motor>=3.3.2
voyageai>=0.1.4
